requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
python-jose==3.3.0
passlib==1.7.4
pytest==8.3.5
httpx==0.28.1
python-multipart==0.0.20
bcrypt==4.2.0